proxy/requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
distro==1.9.0
fastapi==0.110.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
mangum==0.19.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.13.2
uvicorn==0.27.1
